// vanilla/src/enums/filmForm/FilmFormSelectors.ts
/**
 * Enum with the selectors of the elements of film form.
 */
export enum FilmFormSelectors {
  CharactersFieldset = '#characters-container',
  PlanetsFieldset = '#planets-container',
  AddProducerButton = '.add-producer-button',
  RemoveProducerButton = '.remove-producer-button',
  Title = '#title-input',
  OpeningCrawl = '#opening-crawl-input',
  EpisodeId = '#episode-id-input',
  Director = '#director-input',
  ReleaseDate = '#release-date-input',
  ProducerInput = '.producer-input',
  ProducerInputsContainer = '.producer-inputs-container',
}